Code moved into src dir
